+++ command/apply/internal/nginx/nginx.go
@@ -3,9 +3,8 @@ package nginx
 import "fmt"
 
 var conf = `server {
-    listen      8080 default_server;
-    listen      [::]:8080 default_server;
-    server_name _;
+    listen      80;
+    listen      [::]:80;
     set         $base /app;
     root        $base/%s;
 
@@ -15,7 +14,7 @@ var conf = `server {
     fastcgi_read_timeout 240s;
 
     # security
-    include     craftcms/security.conf;
+    include     nitro/security.conf;
 
     # include custom conf files
     include     /app/*nitro.conf;
@@ -29,11 +28,11 @@ var conf = `server {
     }
 
     # additional config
-    include craftcms/general.conf;
+    include nitro/general.conf;
 
     # handle .php
     location ~ \.php$ {
-        include craftcms/php_fastcgi.conf;
+        include nitro/php_fastcgi.conf;
     }
 
     # Allow fpm ping and status from localhost