+++ command/database/backup.go
@@ -76,9 +76,9 @@ func backupCommand(home string, docker client.CommonAPIClient, output terminal.O
 			// create the backup command based on the compatibility type
 			switch compatibility {
 			case "postgres":
-				opts.Commands = []string{"pg_dump", "--username=nitro", db, "-f", "/tmp/" + opts.BackupName}
+				opts.Commands = []string{"pg_dump", "--username=nitro", db, "--clean", "--file", "/tmp/" + opts.BackupName}
 			default:
-				opts.Commands = []string{"mysqldump", "--user=nitro", "-pnitro", db, "--result-file=" + "/tmp/" + opts.BackupName}
+				opts.Commands = []string{"mysqldump", "--user=nitro", "-pnitro", db, "--add-drop-table", "--result-file=" + "/tmp/" + opts.BackupName}
 			}
 
 			output.Pending("creating backup", opts.BackupName)