+++ pkg/caddy/caddy.go
@@ -1,10 +1,9 @@
 package caddy
 
 type UpdateRequest struct {
-	HTTPS   Server `json:"https,omitempty"`
-	HTTP    Server `json:"http,omitempty"`
-	Node    Server `json:"node,omitempty"`
-	NodeAlt Server `json:"node_alt,omitempty"`
+	HTTPS Server `json:"https,omitempty"`
+	HTTP  Server `json:"http,omitempty"`
+	Alt   Server `json:"alt,omitempty"`
 }
 
 type Server struct {